Creates common component for textfields
